Tech Writer review - Update README.md
Really great job! There was not any proofreading errors I could find in just one pass. And the only really content change is just to break up this paragraph a little to help with readability and comprehension. 

Also, I spelled out the acronym RAG. Since LLMs are still relatively new to general developers we stood always spell out acronyms to avoid confusion.

"...In order to do this, InquiryAI uses an AI agent that has two tools available to it: MapQuestionToSchema and GenerateFunction.

The MapQuestionToSchema tool maps the user's question to the graph schema.

The GenerateFunction tool generates a function call based on the user's question and the graph schema. The function is then executed on the graph database, then the result is processed by the LLM service and returned to the user. GenerateFunction utilizes a vector RAG (Retrieval-Augmented Generation) process to suggest the most relevant functions to the user's question. Currently, the vector store used by GenerateFunction is ephemeral, meaning that the available queries will not be persisted when the container is shut off. If the container dies, the desired queries available to the CoPilot service will have to be re-registered."